SpecificPrice::deleteByIdCart - check id_cart input parameter
When this function is called with null/0 value, then all specific
prices with id_cart=0 are deleted. This is obviously wrong.

The fix checks input parameter, and perform delete operation only
if $id_cart is valid positive integer

Closes #933

+++ classes/SpecificPrice.php
@@ -160,6 +160,9 @@ class SpecificPriceCore extends ObjectModel
      */
     public static function deleteByIdCart($idCart, $idProduct = false, $idProductAttribute = false)
     {
+        if (! (int)$idCart) {
+            return false;
+        }
         return Db::getInstance()->delete(
             'specific_price',
             '`id_cart` = '.(int) $idCart.($idProduct ? ' AND `id_product` = '.(int) $idProduct.' AND `id_product_attribute` = '.(int) $idProductAttribute : '')
